Feedback, refactor duplicated class
